requirements.txt
Flask 
tensorflow 
Pillow 
google-generativeai 
python-dotenv 
numpy
flask_cors